gitflow-feature-stash: refactor-controller-name-and-remove-comments
