+++ src/main/java/com/college/hotlittlepigs/controller/ControllerRepository.java
@@ -1,4 +1,4 @@
-package com.college.hotlittlepigs.controllers;
+package com.college.hotlittlepigs.controller;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ControllersRepository extends JpaRepository<Controller, Long> {
+public interface ControllerRepository extends JpaRepository<Controller, Long> {
 
   Page<Controller> findAllByStatus(Boolean status, Pageable pageable);
